requirements.txt
aioboto3==10.1.0
aiobotocore==2.4.0
aiohttp==3.8.3
aioitertools==0.11.0
aiosignal==1.2.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
blis==0.7.8
boto3==1.24.59
botocore==1.27.59
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.0.4
cloudpathlib==0.20.0
confection==0.1.5
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
cymem==2.0.6
daphne==3.0.2
dataclasses-json==0.5.14
datasets==2.6.1
Deprecated==1.2.13
dill==0.3.5.1
distro==1.9.0
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
faiss-cpu==1.7.3
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.10.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==21.3.0
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.4
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langcodes==3.3.0
langgraph==0.2.61
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.51
langsmith==0.2.10
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.25.0
matplotlib==3.6.2
msgpack==1.1.0
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.8
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.10.2
# numpy==1.23.3
numpy==1.26.4
openai==1.59.6
openapi-schema-pydantic==1.2.4
opencv-python==4.11.0.86
orjson==3.10.14
packaging==24.2
pandas==1.5.1
pathlib_abc==0.1.1
pathy==0.11.0
pdf2image
Pillow==9.4.0
preshed==3.0.7
psycopg2-binary==2.9.9
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pypdf2==3.0.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
python-Levenshtein==0.12.2
pytz==2022.4
PyYAML==6.0
redis==4.3.4
regex==2022.9.13
requests==2.28.1
requests-toolbelt==1.0.0
responses==0.18.0
retrying==1.3.4
s3transfer==0.6.0
schedule==1.1.0
scipy==1.10.0
sentence-transformers
service-identity==21.1.0
six==1.16.0
smart-open==5.2.1
sniffio==1.3.1
# spacy==3.8.3
# spacy>=3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.3
SQLAlchemy==1.4.48
sqlparse==0.4.3
srsly==2.5.0
supervisor==4.2.4
tenacity==8.5.0
# thinc==9.1.1
tiktoken==0.8.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1
Twisted==22.8.0
txaio==22.2.1
typer==0.4.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.12
wasabi==0.10.1
wrapt==1.14.1
xxhash==3.2.0
yarl==1.8.1
zipp==3.9.0
zope.interface==5.4.0
openai-whisper==20240930
gunicorn
rank_bm25==0.2.2
python-dotenv
pyaudio==0.2.14
pyttsx3==2.98